% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{a}
\alias{a}
\title{Standardized effect size}
\usage{
a(parameters)
}
\arguments{
\item{parameters}{Parameters specifying the design}
}
\description{
For non-standardized effects, \code{a} returns the standardized effect size
}
