% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{b}
\alias{b}
\title{A simplified version of the function b that is defined in the manuscript}
\usage{
b(parameters, z, n1, lambda1, lambda2)
}
\arguments{
\item{parameters}{Parameters specifying the design}

\item{z}{First stage z value}

\item{n1}{First stage sample size}

\item{lambda1}{Penalization parameter for type I error}

\item{lambda2}{Penalization parameter for type II error}
}
\description{
\code{b} is a help function to compute optimal designs in the Lagrangian framework
}
