% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_ce.R
\name{c_early}
\alias{c_early}
\title{Optimal Stopping Boundaries}
\usage{
c_early(parameters, n1, lambda1, lambda2)
}
\arguments{
\item{parameters}{The parameters (alpha, power, etc.) you want to use for your design}

\item{n1}{The first-stage sample size}

\item{lambda1}{The penalization parameter for type one error}

\item{lambda2}{The penalization parameter for type two error}
}
\value{
A two dimensional vector (c_f, c_e)
}
\description{
\code{c_early} gives the optimal boundaries for early stopping of an adaptive design
}
\details{
The parameter \code{n1} has to be given as well as the penalty parameters \code{lambda_1} and \code{lambda_2}.
The boundaries are computed as roots of the gradient of the score w.r.t. c_f and c_e.
}
