% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_power.R
\name{cond_power}
\alias{cond_power}
\title{Computing conditional power}
\usage{
cond_power(z, d, parameters)
}
\arguments{
\item{z}{Z_1-value of the first stage}

\item{d}{An object of class \code{design}}

\item{parameters}{Parameters secifying the design}
}
\description{
\code{cond_power} computes the conditional power for a given z_1-value.
}
\details{
Note that this function assumes a Gaussian distribution and a t-approximation has not been implemented yet.
}
