% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculation_conditional_power.R
\name{cond_power_design}
\alias{cond_power_design}
\title{Defines design based on conditional power}
\usage{
cond_power_design(parameters, method = c("IN", "BK", "PH"))
}
\description{
Defines design based on conditional power
}
