% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_ce.R
\name{d_cf}
\alias{d_cf}
\title{d Score / d c_f}
\usage{
d_cf(parameters, n1, lambda1, lambda2, x)
}
\arguments{
\item{parameters}{The parameters (alpha, power, etc.) you want to use for your design}

\item{n1}{The first-stage sample size}

\item{lambda1}{The penalization parameter for type one error}

\item{lambda2}{The penalization parameter for type two error}

\item{x}{The z-value of the first stage}
}
\description{
\code{d_cf} gives the gradient of the score w.r.t. c_f.  Mainly needed for \link{c_early}
}
