% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{design}
\alias{design}
\title{Creating design objects}
\usage{
design(cf, ce, n1, n2, c2)
}
\arguments{
\item{cf}{Boundary for stopping for futility}

\item{ce}{Boundary for stopping for efficacy}

\item{n1}{First stage sample size}

\item{n2}{Function of z which gives the stage two sample size}

\item{c2}{Function of z which gives the stage two rejection boundary}
}
\value{
An object of class \code{design}
}
\description{
\code{design} creats an object of class \code{design}.
}
