% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_direct.R
\name{direct_design}
\alias{direct_design}
\title{Find the optimal design via direct optimization}
\usage{
direct_design(parameters)
}
\arguments{
\item{parameters}{Parameters specifying the design.}
}
\value{
An object of class \link{design}.
}
\description{
\code{direct_design} comuptes the optimal design via a spline approach.
}
\details{
The z-space is partionated in a grid and the optimal c_2- and n_2 values are computed on every node.
The function is truncated if z is not element of (c_f,c_e) and inside this interval, a spline function is fitted.
}
