% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_direct_design.R
\name{direct_design_smooth}
\alias{direct_design_smooth}
\title{Compute smooth designs via the direct method}
\usage{
direct_design_smooth(parameters)
}
\arguments{
\item{parameters}{Parameters specifying the design}
}
\value{
An object of class design.
}
\description{
\code{direct_design_smooth} is a different way to compute optimal designs via a spline approach.
}
\details{
This function uses \link{stage_two} to find optimal stage-two functions and optimizes then over the stage one
parameters n1, cf and ce. It is a more suitable way than the one that is used in \link{direct_design} because one only
evaluates the stage two-functions inside the interval (cf,ce) and not on a full grid.
This function yields smoother results than \link{direct_design}, but also takes more time.
}
