% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed_response_direct_optimization.R
\name{dr_design}
\alias{dr_design}
\title{Compute designs for delayed response situation}
\usage{
dr_design(theta, sigma, n1 = NA, lb_n2, alpha, beta)
}
\description{
Compute designs for delayed response situation
}
