% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindLambda.R
\name{err}
\alias{err}
\title{Evaluate the deviation of the desired error constraints}
\usage{
err(parameters, lambda1, lambda2)
}
\arguments{
\item{parameters}{Parameters specifying your design}

\item{lambda1}{Penalization parameter for type I error}

\item{lambda2}{Penalization parameter for type II error}
}
\value{
A two-dimensional vector which gives the deviation of type I resp. type II error.
}
\description{
\code{err} approximates the deviation of a design implemented with \code{lambda1} and \code{lambda2} of the error constraints that
were specified in \code{parameters}.
}
\details{
This is a numerical approxmiation to allow a faster search for the correct lambda-values.
\link{find_lambda} is based on this function.
}
