% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_n.R
\name{exp_n}
\alias{exp_n}
\title{Computing the expected sample size for an effect size}
\usage{
exp_n(effect, d, parameters)
}
\arguments{
\item{effect}{Effect size for which the expected sample size should be calculated}

\item{d}{An object of class \link{design}}

\item{parameters}{Parameters specifying the design}
}
\description{
\code{exp_n} computes the expected sample size for a design under a certain effect size.
}
