% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindLambda.R
\name{find_lambda}
\alias{find_lambda}
\title{Find the correct values of lambda}
\usage{
find_lambda(parameters)
}
\arguments{
\item{parameters}{Parameters specifying the design.}
}
\value{
A two-dimensional vector (lambda1, lambda2).
}
\description{
\code{find_lambda} returns the values of lambda that yield to the desired type I - and type II - error.
}
\details{
Use these values in \link{lagrange_design} to obtain an optimal design with type I - and type II - error rate
as prespecified.
}
