% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindLambda.R
\name{find_lambda_direct}
\alias{find_lambda_direct}
\title{Find lambda values faster}
\usage{
find_lambda_direct(parameters)
}
\arguments{
\item{parameters}{Parameters specifying the design.}
}
\value{
A two-dimensional vector (lambda1, lambda2).
}
\description{
\code{find_lambda_direct} is a way to find approximately correct lambda-values faster than with \link{find_lambda}.
}
\details{
This function is based on \link{direct_design}. Note therefore that is does not yield a purely Lagrangian approach
anymore and is more liable to numerical errors.
}
