% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedDesign.R
\name{fixed}
\alias{fixed}
\title{Compute the fixed design}
\usage{
fixed(parameters)
}
\arguments{
\item{parameters}{Parameters specifying the design}
}
\description{
\code{fixed} computes the fixed design.
}
