% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{int}
\alias{int}
\title{Integral needed for the score}
\usage{
int(parameters, n1, lambda1, lambda2, cf, ce)
}
\arguments{
\item{parameters}{Parameters specifying the design}

\item{n1}{First stage sample size}

\item{lambda1}{Penalization parameter for type I error}

\item{lambda2}{Penalization parameter for type II error}

\item{cf}{Boundary for stopping for futility}

\item{ce}{Boundary for stopping for efficacy}
}
\description{
\code{int} is a help function that is needed for the score in the Lagrangian framework.
}
\details{
See the manuscript for details.
}
