% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_normal.R
\name{inverse_normal_design}
\alias{inverse_normal_design}
\title{Compute optimal sample size based on inverse normal method}
\usage{
inverse_normal_design(parameters)
}
\description{
Compute optimal sample size based on inverse normal method
}
