% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_design.R
\name{jt_design}
\alias{jt_design}
\title{Compute the design by Jennison and Turnbull (2015)}
\usage{
jt_design(theta, sigma, n1 = NA, lb_n2, alpha, beta)
}
\arguments{
\item{theta}{Effect size for which the expected sample size should be minimized}

\item{sigma}{Known standard deviation}

\item{n1}{First stage sample size. Write \code{NA} to obtain the optimal n_1.}

\item{lb_n2}{Lower bound for the stage two sample size due to e.g. delayed response}

\item{alpha}{Maximal type I error rate}

\item{beta}{Maximal type II error rate}
}
\description{
\code{jt_design} computes the design proposed by Jennison and Turnbull (2015).
It is based on the inverse normal combination test.
The n_2-function is termined in order to minimize expected sample size under the alternative.
}
\details{
Note that the sample sizes must be given per group. The total sample size is 2*(n_1 + n_2).
}
