% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_lagrange.R
\name{lagrange_design}
\alias{lagrange_design}
\title{Find the optimal Lagrangian design}
\usage{
lagrange_design(parameters, lambda1, lambda2)
}
\arguments{
\item{parameters}{The parameters (alpha, power, etc.) you want to use for your design}

\item{lambda1}{Penalization parameter for type I error}

\item{lambda2}{Penalization parameter for type II error}
}
\value{
An object of class \code{design}
}
\description{
Compute the optimal design for predefined Lagrangian penalty terms.
}
\details{
Note that this design does not hold type I or type II error constraints necessarily.
To find values of lambda1 and lambda2 which yield the exact error constraints use \link{find_lambda}.
}
