% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed_response_lagrange.R
\name{lagrange_dr_design}
\alias{lagrange_dr_design}
\title{Optimal design in delayed response situation}
\usage{
lagrange_dr_design(theta, sigma, n1 = NA, lb_n2, alpha, beta)
}
\arguments{
\item{theta}{Expeted effect under the alternative}

\item{sigma}{Standard deviation}

\item{n1}{First-stage sample size}

\item{lb_n2}{Lower limit for stage-two sample size}

\item{alpha}{Type one error rate}

\item{beta}{Type two error rate}
}
\value{
An object of class design
}
\description{
Optimal design in delayed response situation
}
