% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1.R
\name{n1}
\alias{n1}
\title{Find the optimal stage one sample size}
\usage{
n1(parameters, lambda1, lambda2)
}
\arguments{
\item{parameters}{The parameters (alpha, power, standardized effect) whith which you want to build your design}

\item{lambda1, lambda2}{The Lagrange penalization parameters}
}
\description{
\code{n1} computes the stage one sample size of an optimal adaptive design
}
