% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2.R
\name{n2}
\alias{n2}
\title{Compute n_2 function}
\usage{
n2(parameters, z, n1, lambda1, lambda2, cf, ce)
}
\description{
Computes the optimal n_2-value for a given first stage z_1-value in the Lagrangian framework.
}
\details{
\code{n2} differs from \link{response} only so far that the values are set to 0 outside the interval (cf,ce).

@param parameters Parameters specifying the design
 @param z Z_1-value of the first stage
 @param n1 First stage sample size
 @param lambda1 Penalization parameter for type I error
 @param lambda2 Penalization parameter for type II error
 @param cf Boundary for stopping for futility
 @param ce Boundary for stopping for efficacy
}
