% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{nodes}
\alias{nodes}
\title{Define equi distance nodes}
\usage{
nodes(l, u, N)
}
\arguments{
\item{l}{Lower Bound}

\item{u}{Uper Bound}

\item{N}{In total, there are 4N+1 nodes}
}
\description{
\code{nodes} defines equi distance nodes inside the invterval (l,u)
}
