% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_alpha.R
\name{opt_alpha}
\alias{opt_alpha}
\title{Compute the type I error rate of a design}
\usage{
opt_alpha(d)
}
\arguments{
\item{d}{An object of class \code{design}.}
}
\description{
\code{opt_alpha} gives the type I error rate of a design.
}
