% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimal_design}
\alias{optimal_design}
\title{Compute an optimal adaptive design}
\usage{
optimal_design(effect, alpha, pow, effect_null = 0, sd = 1,
  standardized = T, t_approx = F, lagrange = T)
}
\arguments{
\item{effect}{The effect on that the power is computed}

\item{alpha}{The maximal type I error rate}

\item{pow}{The minimal power}

\item{effect_null}{The effect under the null hypothesis. Default is 0.}

\item{sd}{The standard deviation. Default is 1.}

\item{standardized}{Logical. Is \code{effect} already standardized? Default is \code{T}.}

\item{t_approx}{Logical. Should a t-approximation be used? Default is \code{F}.}

\item{lagrange}{Logical. Should a Lagrangian procedure be used? Default is \code{T}.}
}
\value{
An object of class \link{design}
}
\description{
\code{optimal_design} returns an optimal two-stage adaptive design.
}
\details{
The optimality criterion is given as the expected sample size under the null hypothesis.
If \code{standardized=T}, the values of \code{effect_null} and {sd} are set to their default, even if they were specified.

The use of a t-approximation (\code{t_approx=T}) or a Lagrangian approach (\code{lagrange=T}) yields more precise results,
but makes the calculation slowlier.

Note that there is no Lagrangian implementation of the t_approxmation. Hence, the value of \code{lagrange} will be
ignored, if \code{t_approx} is set to \code{T}.
}
