% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalGSD.R
\name{optimal_gsd}
\alias{optimal_gsd}
\title{Compute the optimal group sequential design}
\usage{
optimal_gsd(parameters)
}
\arguments{
\item{parameters}{Parameters specifying the design}
}
\description{
\code{optimal_gsd} computes the group sequential design that is optimal w.r.t. expected sample size under the
alternative hypothesis and holdes presepcified type I and type II error constraints.
}
