% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_normal_optimal.R
\name{optimal_inverse_normal_design}
\alias{optimal_inverse_normal_design}
\title{Compute the optimal design based on the inverse normal combination test}
\usage{
optimal_inverse_normal_design(parameters)
}
\description{
Compute the optimal design based on the inverse normal combination test
}
