% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{parameters}
\alias{parameters}
\title{Creating parameters object}
\usage{
parameters(mu, alpha, beta)
}
\arguments{
\item{mu}{Standardized effect value of the alternative hypothesis}

\item{alpha}{Maximal type I error rate}

\item{beta}{Maximal type II error rate}
}
\value{
An object of class \code{parameters.}
}
\description{
\code{parameters} creats an object of class \code{parameters} which is needed for many functions in this package.
}
