% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2.R
\name{plot_c2}
\alias{plot_c2}
\title{Plot the c2-function}
\usage{
plot_c2(d)
}
\arguments{
\item{d}{An object of class \link{design}}
}
\description{
\code{plot_c2} plots the c_2 function of a design in dependence of the first stage z-value.
}
