% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_power.R
\name{plot_cond_power}
\alias{plot_cond_power}
\title{Plotting conditional power}
\usage{
plot_cond_power(d, parameters)
}
\arguments{
\item{d}{An object of class \code{design}}

\item{parameters}{Parameters secifying the design}
}
\description{
\code{cond_power} plots the conditional power curve of a design.
}
\details{
Note that this function assumes a Gaussian distribution and a t-approximation has not been implemented yet.
}
