% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_n.R
\name{plot_exp_n}
\alias{plot_exp_n}
\title{Plotting expected sample size}
\usage{
plot_exp_n(d, parameters)
}
\arguments{
\item{d}{An object of class \link{design}}

\item{parameters}{Parameters specifying the design}
}
\description{
\code{plot_exp_n} plots the expected sample size for different effect sizes
}
