% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_n.R
\name{plot_n}
\alias{plot_n}
\title{Plotting total sample size}
\usage{
plot_n(d, parameters)
}
\arguments{
\item{d}{An object of class \link{design}}

\item{parameters}{Parameters specifying the design}
}
\description{
\code{plot_n} plots the total sample size of a design dependent on the z-value of the first stage.
}
