% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2.R
\name{plot_n2}
\alias{plot_n2}
\title{Plot the n2 function}
\usage{
plot_n2(d)
}
\arguments{
\item{d}{An object of class \link{design}.}
}
\description{
\code{plot_n2} plots the n2-function of a design.
}
