% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{plot_power}
\alias{plot_power}
\title{Plot power dependent on true effect}
\usage{
plot_power(d, parameters)
}
\arguments{
\item{d}{An object of class \code{design}.}

\item{parameters}{Parameters speficying the situation. See \link{parameters} for details.}
}
\description{
\code{plot_power} plots the power curve for the true effect.
}
\details{
The true effect size has to be specified in \code{parameters} as well as type I - and type II - error constraints.
}
