% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{real_power}
\alias{real_power}
\title{Computing the conditional power for differnt true effect sizes}
\usage{
real_power(effect, d, parameters)
}
\arguments{
\item{effect}{Effect size on which one wants to calculate the power.}

\item{d}{An object of class \code{design}.}

\item{parameters}{Parameters speficying the situation. See \link{parameters} for details.}
}
\description{
\code{real_power} gives the power of a design at a specified effect size
}
\details{
In \code{parameters} one has to specify type I-, type II-error and effect size.
}
