% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2.R
\name{response}
\alias{response}
\title{Compute concrete n_2-value}
\usage{
response(parameters, n1, lambda1, lambda2, z)
}
\description{
\code{response} computes the concrete n_2-value for a given first stage z_1-value.
}
\details{
Here, the Lagrangian framework is used du to the fact that the computation of n_2 can be reduced to a node-finding problem.

@param parameters Parameters specifying the design
 @param n1 First stage sample size
 @param lambda1 Penalization parameter for type I error
 @param lambda2 Penalization parameter for type II error
 @param z Z_1-value of the first stage
 @export
}
