% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score in the Lagrangian framework}
\usage{
score(parameters, cf, ce, n1, lambda1, lambda2)
}
\arguments{
\item{parameters}{Parameters specifying the design}

\item{cf}{Boundary for stopping for futility}

\item{ce}{Boundary for stopping for efficacy}

\item{n1}{First stage sample size}

\item{lambda1}{Penalization parameter for type I error}

\item{lambda2}{Penalization parameter for type II error}
}
\description{
\code{score} defines the score in the Lagrangian framework.
}
\details{
The score is given as the expected sample size of the trial under the alternative hypothesis.
}
