% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageTwo.R
\name{score_smooth}
\alias{score_smooth}
\title{Smooth score version}
\usage{
score_smooth(parameters, cf, ce, n1)
}
\arguments{
\item{cf}{Boundary for stopping for futility}

\item{ce}{Boundary for stopping for efficacy}

\item{n1}{Stage one sample size}

\item{Parameters}{Parameters specifying the design}
}
\description{
\code{score_smooth} gives the version of the score that is needed for \link{score_direct_smooth}.
}
