% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageTwo.R
\name{stage_two}
\alias{stage_two}
\title{Compute optimal stage two values for given first stage}
\usage{
stage_two(parameters, cf, ce, n1)
}
\arguments{
\item{parameters}{Parameters specifying the design}

\item{cf}{Boundary for stopping for futility}

\item{ce}{Boundary for stopping for efficacy}

\item{n1}{First stage sample size}
}
\value{
A vector. The first half give the c_2, and the second the n_2-values, on a equidistance grid inside the
interval (cf,ce).
}
\description{
\code{stage_two} computes the functions c_2 and n_2 that hold the error constraints and are optimal w.r.t.
expected sample size under the alternative for a prespecified first stage.
}
\details{
\code{stage_two} is the base of \link{direct_design_smooth}.
}
