% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_optimization.R
\name{t_design}
\alias{t_design}
\title{Find the optimal design when using a t-approximation}
\usage{
t_design(parameters)
}
\arguments{
\item{parameters}{Parameters specifying the design}
}
\value{
An object of class \link{design}
}
\description{
\code{t_design} computes the optimal design when a t-approximation is used.
}
