% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_Score.R
\name{t_score}
\alias{t_score}
\title{Version of the score when using t-approximation}
\usage{
t_score(parameters, cf, ce, c2, n1, n2, w)
}
\arguments{
\item{parameters}{Parameters specifying the design}

\item{cf}{Boundary for stopping for futility}

\item{ce}{Boundary for stopping for efficacy}

\item{c2}{c_2 values on the nodes}

\item{n1}{Stage one sample size}

\item{n2}{n_2 values on the nodes}

\item{w}{nodes}
}
\description{
\code{t_score} gives the version of the score when the t_approximation is used.
}
\details{
Mainly needed for \link{t_design}
}
