% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageTwo.R
\name{type_one_smooth}
\alias{type_one_smooth}
\title{Type I version for smooth direct designs}
\usage{
type_one_smooth(parameters, cf, c2, h, N, w)
}
\arguments{
\item{parameters}{Parameters specifying the design}

\item{cf}{Boundary for stopping for futility}

\item{c2}{c_2-values}

\item{h}{Distance between two nodes}

\item{N}{4N+1 gives the number of nodes}

\item{w}{nodes inside the interval (cf,ce)}
}
\description{
\code{type_one_smooth} gives the version of the type I error that is needed for \link{stage_two}.
}
