% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypeTwo.R
\name{type_two}
\alias{type_two}
\title{Compute the type II error}
\usage{
type_two(parameters, cf, ce, nodes, c2, n1, n2)
}
\arguments{
\item{parameters}{Parameters (alpha, power, standardized effect) whith which you want to build your design}

\item{cf}{Boundary for stopping for futility after the first stage}

\item{ce}{Boundary for stopping for efficacy after the first stage}

\item{nodes}{Nodes for the stage two functions}

\item{c2}{c_2-values that correspond to \code{nodes}}

\item{n1}{First stage sample size}

\item{n2}{n_2-values that correspond to \code{nodes}}
}
\description{
\code{type_two} computes the type II error approximately if the \code{c_2}-function and the \code{n_2}-function
are approximated via a set of splines.
}
\details{
Note that nodes, c2 and n2 have to be of the same length.
}
