% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageTwo.R
\name{type_two_smooth}
\alias{type_two_smooth}
\title{Type II version for smooth direct designs}
\usage{
type_two_smooth(parameters, cf, c2, n1, n2, h, N, w)
}
\arguments{
\item{parameters}{Parameters specifying the design}

\item{cf}{Boundary for stopping for futility}

\item{c2}{c_2-values}

\item{n1}{First stage sample size}

\item{n2}{n_2-values}

\item{h}{Distance between two nodes}

\item{N}{4N+1 gives the number of nodes}

\item{w}{nodes inside the interval (cf,ce)}
}
\description{
\code{type_two_smooth} gives the version of the type II error that is needed for \link{stage_two}.
}
