% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{parameters}
\alias{parameters}
\title{Creating parameters object}
\usage{
parameters(mualt, alpha, beta, mu0 = 0, sigma = 1)
}
\arguments{
\item{mualt}{Effect value of the alternative hypothesis}

\item{alpha}{Maximal type I error rate}

\item{beta}{Maximal type II error rate}

\item{mu0}{Effect value of the null hypothesis. Default value is 0.}

\item{sigma}{Standard deviation. Default value is 1.}
}
\value{
An object of class \code{parameters.}
}
\description{
\code{parameters} creats an object of class \code{parameters} which is needed for many functions in this package.
}
\details{
If \code{mu0} and \code{sigma} are not speficied, \code{mualt} is assumed to be the standardized effect size
of the alternative.
}
